requirements.txt
streamlit==1.8.1
pandas==1.3.4
numpy==1.20.3
matplotlib==3.4.1
Pillow==8.4.0
beautifulsoup4==4.10.0
requests==2.25.1
